+++ CHANGELOG.md
@@ -13,6 +13,16 @@ Version 2.0.0
 - Added plugins directory to the views path.  This makes it possible for
   plugins to render their own ejs views or pages in their own folders by
   using paths of the format "pluginName/path/to/page-or-view".
+- Replaced the per instance copy of the shared Factorio mods with
+  symlinks.  On Windows hard links are used instead due to the
+  privileges requirements of symlinks.
+- Changed the per instance scenario folder to be linked instead of
+  copied on instance creation.
+- Instance id is no longer derived from the rcon password, instead it's
+  generated upon instance creation and stored in the instance config.
+- The game port, RCON port and RCON password instance config entries are are
+  now null by default, indicating that a random one will be generated every
+  time the instance is started.
 
 ### Breaking Changes
 
@@ -25,6 +35,9 @@ Version 2.0.0
 - Removed config management from the command line and the server manager.
 - Moved ejs templates into views folder and changed their extension to
   .ejs.  Breaks playerManager.
+- Mods are no longer copied from the per instance instanceMods directory.
+  If you need per instance mods you can now place them directly in the mods
+  directory inside the instance folder.
 
 
 Version 1.2.2